Create LICENSE suggestion
Can I suggest adding a license to your awesome work?
The most common license I see on Github is the MIT one and that is why I suggested it.
Of course feel free to change it to the license you want. Also, I allowed myself to specify the name I found in setup.py (Lai Kit So). Do not forget to change it if it is not appropriate.
The license file is just a kind reminder because unless there is a license included to the project, we are not licensed to use your awesome widget.